Split watch
